feat: Implement BCF serialization for GL renderer
Replace broken voxel grid sampling with Binary Cube Format (BCF) serialization
for WebGL 2.0 octree rendering. This fixes the GL renderer which was producing
black/empty output due to coordinate system mismatch.

Major changes:
- Change all renderer types to use Cube<u8> directly (no i32 conversion)
- Remove broken voxel grid sampling (sample_cube_at_position, create_octree_texture)
- Add BCF serialization in GlTracerGl::new()
- Upload BCF data as 1D-like 2D texture (R8UI format, width×1 height)
- Rewrite fragment shader for BCF traversal:
  - Add BCF byte reading functions (readU8, readPointer, decodeTypeByte)
  - Implement BCF node parsing (parseBcfNode, getOctant)
  - Add stack-based octree traversal (raycastBcfOctree)
  - Replace old 3D texture sampling with direct BCF parsing

Technical details:
- BCF data stored as usampler2D texture on GPU
- Stack-based iterative traversal (no recursion in GLSL)
- Supports all BCF node types (inline leaf, extended leaf, octa-with-leaves, octa-with-pointers)
- World space coordinates [-1, 1]³ (matches Rust octree coordinate system)

Files modified:
- crates/renderer/src/gl_tracer.rs: BCF upload, uniform binding
- crates/renderer/src/shaders/octree_raycast.frag: Complete shader rewrite
- crates/renderer/src/cpu_tracer.rs: Update to Cube<u8>
- crates/renderer/src/gpu_tracer.rs: Update to Cube<u8>
- crates/renderer/src/scenes/octa_cube.rs: Return Cube<u8>

+++ crates/renderer/src/gpu_tracer.rs
@@ -46,7 +46,7 @@ void main() {
 
 /// Compute shader raytracer
 pub struct GpuTracer {
-    cube: Rc<Cube<i32>>,
+    cube: Rc<Cube<u8>>,
     #[allow(dead_code)]
     bounds: CubeBounds,
     gl_state: Option<GpuTracerGl>,
@@ -69,7 +69,7 @@ struct GpuTracerGl {
 }
 
 impl GpuTracer {
-    pub fn new(cube: Rc<Cube<i32>>) -> Self {
+    pub fn new(cube: Rc<Cube<u8>>) -> Self {
         Self {
             cube,
             bounds: CubeBounds::default(),
@@ -139,7 +139,7 @@ impl GpuTracer {
     }
 
     /// Get reference to the cube
-    pub fn cube(&self) -> &Rc<Cube<i32>> {
+    pub fn cube(&self) -> &Rc<Cube<u8>> {
         &self.cube
     }
 
@@ -151,7 +151,7 @@ impl GpuTracer {
         pos: glam::Vec3,
         dir: glam::Vec3,
         _max_depth: u32,
-    ) -> Result<Option<cube::Hit<i32>>, String> {
+    ) -> Result<Option<cube::Hit<u8>>, String> {
         Ok(cube::raycast(&self.cube, pos, dir, None))
     }
